object_and_method.cpp
Write a CPP program, where you have 2 data items (both int) and one method in a class. Do not give any access specifier for both data and method. Now write main() and create an object and try to initialise the data directly and also try to invoke the method directly. Analyse this and note down findings.