Create ClassWithConstructorAndPrintFunction.cpp
A C++ program with a class that includes a constructor. The class should have two integer members and one float member. Initialize these members in the constructor. Then, create another function inside the same class to print the values. Finally, create an object of this class and use the function to print the values.